requirements.txt
streamlit 
openai 
requests 
feedparser 
fpdf 
deep-translator 
python-dotenv    